Remove useless check in kb creation test (#2)
We were checking if after creation the list of KB had a length of 2, while after the addition the list should contain 3 KB. This check was passing because the creation is a bit slow and also because we use to start with 1 KB (while we now have 2 kbs).
In this PR we remove the wrong check. What is important is  the newly created KB appears in the list after creation, not how many kbs there are in the list. This test is now more robust and future-proof (we can add more KBs without risking breaking this test).