Prepare GH Action configuration for running e2e in prod (#58)
This PR is just a first step of splitting the GH action configuration in two, so we can run e2e on prod. Once this PR will be merged, we will create a branch `e2e-on-prod` which will have specific cypress configuration for prod, as well as tests on US zone.