Wait for citation block to appear AFTER getting the answer (#85)
We had a 10s timeout to get the answer, but actually it's the citation block which is slow to be displayed after we get the answer itself. So we put back the same 8s to get answer (as in the other test) and we add a 6s timeout on getting citations afterward.