Avoid empty POST to stop tasks
We had a problem when senting an empty POST triggered a 411 response
from Google's load balancer. This is a known issue where the LB is
deviating slightly from the standards (RFC7230, 3.3.3). So we send
a semantically empty body but with actual length in order to make
sure a Content-Length header is sent.

This mimics how frontend does the query, sending an empty JSON
model, so we get E2E that are closer to real world usage.